Change default of "syntax highlight" option to "long", Add basic syntax highlight styles for LaTeX and HTML.
git-svn-id: svn://svn.code.sf.net/p/docutils/code/trunk@7514 929543f6-e4f2-0310-98a6-ba3bd3dd1d04

+++ docutils/docutils/parsers/rst/__init__.py
@@ -136,7 +136,7 @@ class Parser(docutils.parsers.Parser):
           '"long", "short", or "none (no parsing)". Default is "short".',
           ['--syntax-highlight'],
           {'choices': ['long', 'short', 'none'],
-           'default': 'short', 'metavar': '<format>'}),))
+           'default': 'long', 'metavar': '<format>'}),))
 
     config_section = 'restructuredtext parser'
     config_section_dependencies = ('parsers',)
